Split addToQueueService
